Refactor NGFF utilities and robustify writers
Move NGFF/Zarr helper logic into utils/ngff.py and clean up related modules. Introduce a ZarrWriteConfig dataclass and centralize format resolution, compressor builders, chunk handling, pyramid writers, axes/coordinate/OMERO helpers, and unit/color normalization in ngff.py. Update to_zarr.py to import these helpers (removing duplicated implementations). Tighten create_empty_group to infer/validate root ngff and zarr_format, propagate compressor settings, and set v2 chunk_key_encoding. Improve write_image_region and write_label_region with safer group resolution (nested paths), clearer warnings, shape/dimension checks, proper dataset handling, corrected index scaling logic, and flush safety. Remove deprecated add_label.py and load_label.py and add pymif.zip binary. Adjust zarr_manager import to use the new ngff.ZarrWriteConfig/_resolve_format.
