+++ pymif/microscope_manager/utils/create_empty_dataset.py
@@ -1,18 +1,16 @@
 from typing import Dict, Any
 import zarr
 
-from .to_zarr import (
+from .ngff import (
+    _set_group_ngff_metadata,
+    _build_v2_compressor,
+    _build_v3_compressors,
     ZarrWriteConfig,
     _resolve_format,
     _build_axes,
     _build_coordinate_transformations,
     _build_omero_metadata,
-    _build_v2_compressor,
-    _build_v3_compressors,
 )
-
-from .ngff import _set_group_ngff_metadata
-
 
 def create_empty_dataset(
     root: zarr.Group,
